Lemmas about ContPL programs; another naturality principle
Principles for transforming (map h) * (a ContPL program) into another
ContPL program. Not sure where these really belong, so they're in
Samplers.v for now. The "fail" at the end is due, I think, to some
complexity of notation. Also, added to the spec that join, as well as
ret, is a natural transformation.