pstream axiom, first stab
This is very ugly, and has a *lot* of issues with the "Coq can't infer
something which should be obvious" problem, but I believe that this
proposition correctly expresses what we want of pstream. (I changed
pstream to a more general version, similar to the more general stream
constructor.) Possibly there is an easier way to do this with a program.
(The change to Category.v is a trivial renaming.)