Third monad law
Major theorem "Bind_Bind", corresponding to the monad 'associativity'.
Intent is to translate this into a more program-looking version, which
will say something like the equivalence of (x <- (y <- A; f y); g x)
with (y <- A; x <- f y; g x). This is necessary (I think) for the
bind_sampler.