Change sampler proofs to make less point-free
Make them less point-free (in particular, most have their computational
content in the form of a contPL program). This has had the happy
side-effect of making the proofs resemble each other more. This applies
to the sampler proofs which have been completed at this time, which are:
constant_sampler, infinite_sampler, and pullback_sampler.