wait for heads, initial version
The first partial sampler, wait_for_heads, is more or less complete.
There are a lot of holes in the things that lead up to it, but I think
the logical structure of the proof is basically there. Also, I changed
unit_uniq.

Holes:
- The inclusion [dom f] -> Cantor is an embedding.
- The following is an embedding:
(cons snd fst ∘ copair ⟨ id, Zero ⟩ ⟨ pb_fst f strict, One ⟩)
(I suspect that this is essentially a composition of
cons-Zero, cons-One, and pb_fst f strict being embeddings.
(That last map is the first bullet point.)
- (cons Zero)*infinite_coinflips = infinite_coinflips/2
- cons-One is an embedding
- cons-Zero is an embedding
(right now this is in the code as following from point 2,
but the actual proof of point 2 will probably involve proving
it directly)
- (cons One)*infinite_coinflips = infinite_coinflips/2
- the sum of 1/2^(n+1) is equal to 1.

Holes in toy examples:
- the fixpoint example in Spec/Prob is now totally broken
- for the series example in Sums, there's a hole that's basically
1 + (1/2) * (1 + 1) = 1 + 1. maybe there is a way to do this
automatically?
some of the +'s are addition of Qnns, some of them are of LRnns. :)